Continue build refactoring: move massive tasks configurations to a separate file, enhance build files structure
